feat:0004 os and tensor fixed.

+++ pcdet/ops/iou3d_nms/iou3d_nms_utils.py
@@ -130,9 +130,9 @@ def nms_gpu(boxes, scores, thresh, pre_maxsize=None, **kwargs):
         order = order[:pre_maxsize]
 
     boxes = boxes[order].contiguous()
-    keep = torch.LongTensor(boxes.size(0))
+    keep = torch.IntTensor(boxes.size(0))
     num_out = iou3d_nms_cuda.nms_gpu(boxes, keep, thresh)
-    return order[keep[:num_out].cuda()].contiguous(), None
+    return order[keep[:num_out].long().cuda()].contiguous(), None
 
 
 def nms_normal_gpu(boxes, scores, thresh, **kwargs):
